Moving Objects BP and Character Interaction setup
lot of stuff done!
I set up the definitive system to pick up stuff and handle
activators/triggers as levers, buttons and switches and this kind of
stuff.
All movable objects (the books for example) will use now the same BP
class: MovableObj.
I set it up so that it behaves exactly as picked up objects already
behaved, but now when placing it on the editor you can select the mesh
you want it to be. So is a single BP for all objects, which is kinda
powerful I think. We could modify it in many ways, for example using
environmental sounds when they hit the ground or another surface, and
stuff like that.