LeverBP and CharacterBP adjustments. Livingroom.
Set the levers to have two switch positions. changing the Trigger
interface accordingly to have two events (named them Positive and
Negative Triggers).
Slightly changed how the interp disance for the pickup point works,
solving an issue with the pickup location to automatically set to 320
units of distance after a few seconds after pickup, and ignoring
scrollwheel. Not it works.
I found out a basic problem with the pickup mechanic which we will need
to pay attention to: all objects that allow items to be "trapped" need
to be set to WorldStatic, or the Sphere Trace will never be able to
catch a movable object which is inside another mesh (like the piano or
the sink).